fixed and adjusted tests

+++ src/main/java/com/spring/demo/services/UserService.java
@@ -2,10 +2,10 @@ package com.spring.demo.services;
 
 import com.spring.demo.dto.UserDto;
 import com.spring.demo.enums.Role;
+import com.spring.demo.exceptions.UserAlreadyExistsException;
 import com.spring.demo.exceptions.UserNotFoundException;
 import com.spring.demo.persistence.entities.UserEntity;
 import com.spring.demo.persistence.repositories.UserRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 
@@ -16,13 +16,15 @@ import java.util.stream.Collectors;
 @Service
 public class UserService {
 
-	@Autowired
-	private UserRepository userRepository;
+	private final UserRepository userRepository;
+
+	public UserService(UserRepository userRepository) {
+		this.userRepository = userRepository;
+	}
 
 	@Cacheable("users")
 	public List<UserDto> getAllUsers() {
-		List<UserEntity> all = userRepository.findAll();
-		return all
+		return userRepository.findAll()
 			.stream()
 			.map(UserMapper::convertToUserDto)
 			.collect(Collectors.toList());
@@ -42,15 +44,31 @@ public class UserService {
 
 	public UserDto createNewUser(UserDto userDto) {
 		UserEntity userEntity = UserMapper.convertToUserEntity(userDto);
+		if (userRepository.findByUsername(userEntity.getUsername()).isPresent()) {
+			throw new UserAlreadyExistsException(userEntity.getUsername());
+		}
 		userRepository.saveAndFlush(userEntity);
 		return UserMapper.convertToUserDto(userEntity);
 	}
 
-	public void deleteUserById(int userId) {
-		userRepository.deleteUserEntityByUserId(userId);
+	public UserDto updateUser(int userId, UserDto userDto) {
+		var userEntity = userRepository.findById(userId).orElseThrow(() -> new UserNotFoundException(userId));
+
+		if (userEntity.getUsername().equals(userDto.getUsername())) {
+			throw new UserAlreadyExistsException(userDto.getUsername());
+		} else if (userEntity.getEmail().equals(userDto.getEmail())) {
+			throw new UserAlreadyExistsException(userDto.getEmail());
+		} else {
+			userEntity = userRepository.saveAndFlush(UserMapper.convertToUserEntity(userDto));
+		}
+		return UserMapper.convertToUserDto(userEntity);
 	}
 
-	public void deleteUserByUsername(String userName) {
-		userRepository.deleteUsersEntityByUsername(userName);
+	public void deleteUserById(int userId) {
+		if (userRepository.existsById(userId)) {
+			userRepository.deleteUserEntityByUserId(userId);
+		} else {
+			throw new UserNotFoundException(userId);
+		}
 	}
 }
